cleanup for using event listener
