+++ src/plugin/src/main/java/com/defenseunicorns/uds/keycloak/plugin/CustomEventListenerProvider.java
@@ -1,5 +1,9 @@
 package com.defenseunicorns.uds.keycloak.plugin;
 
+import java.nio.charset.StandardCharsets;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 import org.keycloak.events.Event;
 import org.keycloak.events.EventListenerProvider;
 import org.keycloak.events.EventType;
@@ -27,17 +31,42 @@ public class CustomEventListenerProvider implements EventListenerProvider {
             RealmModel realm = this.model.getRealm(event.getRealmId());
             UserModel newRegisteredUser = this.session.users().getUserById(realm, event.getUserId());
 
-            newRegisteredUser.setSingleAttribute("newAttribute", "newAttributeValue");
+            generateMattermostId(newRegisteredUser);
         }
     }
 
     @Override
     public void onEvent(AdminEvent adminEvent, boolean b) {
-
+        // no implementation needed
     }
 
     @Override
     public void close() {
+        // no implementation needed
+    }
 
+    /**
+     * Add a custom user attribute (mattermostid) to enable direct mattermost <>
+     * keycloak auth on mattermost teams edition.
+     *
+     * @param user     the Keycloak user object
+     */
+    private static void generateMattermostId(UserModel user) {
+        if (user != null && user.getEmail() != null){
+            String email = user.getEmail();
+
+            byte[] encodedEmail;
+            int emailByteTotal = 0;
+            Date today = new Date();
+    
+            encodedEmail = email.getBytes(StandardCharsets.US_ASCII);
+            for (byte b : encodedEmail) {
+                emailByteTotal += b;
+            }
+    
+            SimpleDateFormat formatDate = new SimpleDateFormat("yyDHmsS");
+    
+            user.setSingleAttribute("mattermostid", formatDate.format(today) + emailByteTotal);
+        }
     }
 }
