+++ src/plugin/src/main/resources/META-INF/services/org.keycloak.authentication.RequiredActionFactory
@@ -1,2 +1,1 @@
 com.defenseunicorns.uds.keycloak.plugin.UpdateX509
-com.defenseunicorns.uds.keycloak.plugin.CustomUserAttributes
