+++ src/realm.json
@@ -1553,7 +1553,8 @@
     "eventsEnabled": true,
     "eventsExpiration": 1800,
     "eventsListeners": [
-        "jboss-logging"
+        "jboss-logging",
+        "custom-registration-listener"
     ],
     "enabledEventTypes": [
         "SEND_RESET_PASSWORD",
@@ -2712,15 +2713,6 @@
             "enabled": true,
             "defaultAction": false,
             "priority": 1001,
-            "config": {}
-        },
-        {
-            "alias": "CUSTOM_USER_ATTRIBUTES",
-            "name": "Custom User Attributes",
-            "providerId": "CUSTOM_USER_ATTRIBUTES",
-            "enabled": true,
-            "defaultAction": true,
-            "priority": 1,
             "config": {}
         }
     ],
